+++ softplc/src/test/java/de/peteral/softplc/datatype/DataTypeFactoryTest.java
@@ -1,7 +1,6 @@
 package de.peteral.softplc.datatype;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 import java.util.Arrays;
 
@@ -16,86 +15,102 @@ import de.peteral.softplc.address.ParsedAddress;
 
 @SuppressWarnings({ "javadoc" })
 @RunWith(Parameterized.class)
-public class DataTypeFactoryTest {
+public class DataTypeFactoryTest
+{
 
-	/* @formatter:off */
+    /* @formatter:off */
 	@Parameters(name = "{index}: {0} = {1}")
 	public static Iterable<Object[]> getParameters() {
 		return Arrays.asList(new Object[][] {//
 				{ new ParsedAddress("DB100,C10:20"), "Hallo", new byte[] { 'H', 'a', 'l', 'l', 'o', 0x00 }, 20, 1, 0, 1 },
 				{ new ParsedAddress("DB100,C10:10,2"), new String[] { "Hallo", "abcd" }, new byte[] { 'H', 'a', 'l', 'l', 'o', 0x00, 0x00, 0x00, 0x00, 0x00, 'a', 'b', 'c', 'd', 0x00 }, 20, 1, 0, 2 },
-				{ new ParsedAddress("M100,B20"), 20.0, new byte[] { 20 }, 1, 1, 0, 1 },
-				{ new ParsedAddress("M100,B20"), 250.0, new byte[] { (byte) 250 }, 1, 1, 0, 1 },
-				{ new ParsedAddress("M100,B20,2"), new Double[] { 10.0, 20.0 }, new byte[] { 10, 20 }, 2, 1, 0, 2 },
-				{ new ParsedAddress("M100,W20"), 20.0, new byte[] { 0, 20 }, 2, 2, 0, 1 },
-				{ new ParsedAddress("M100,W20"), 40000.0, new byte[] { (byte) 156, (byte) 220 }, 2, 2, 0, 1 },
-				{ new ParsedAddress("M100,W20,2"), new Double[] { 10.0, 20.0 }, new byte[] { 0, 10, 0, 20 }, 4, 2, 0, 2 },
+				{ new ParsedAddress("M100,B20"), 20, new byte[] { 20 }, 1, 1, 0, 1 },
+				{ new ParsedAddress("M100,B20"), 250, new byte[] { (byte) 250 }, 1, 1, 0, 1 },
+				{ new ParsedAddress("M100,B20,2"), new Integer[] { 10, 20 }, new byte[] { 10, 20 }, 2, 1, 0, 2 },
+				{ new ParsedAddress("M100,W20"), 20, new byte[] { 0, 20 }, 2, 2, 0, 1 },
+				{ new ParsedAddress("M100,W20"), 40000, new byte[] { (byte) 156, (byte) 220 }, 2, 2, 0, 1 },
+				{ new ParsedAddress("M100,W20,2"), new Integer[] { 10, 20 }, new byte[] { 0, 10, 0, 20 }, 4, 2, 0, 2 },
 		});
 	}
 	/* @formatter:on */
 
-	public DataTypeFactoryTest(ParsedAddress address, Object value,
-			byte[] bytes, int byteArraySize, int elementSize, int headerSize,
-			int count) {
-		this.address = address;
-		this.value = value;
-		this.bytes = bytes;
-		this.byteArraySize = byteArraySize;
-		this.elementSize = elementSize;
-		this.headerSize = headerSize;
-		this.count = count;
-	}
+    public DataTypeFactoryTest(ParsedAddress address,
+                               Object value,
+                               byte[] bytes,
+                               int byteArraySize,
+                               int elementSize,
+                               int headerSize,
+                               int count)
+    {
+        this.address = address;
+        this.value = value;
+        this.bytes = bytes;
+        this.byteArraySize = byteArraySize;
+        this.elementSize = elementSize;
+        this.headerSize = headerSize;
+        this.count = count;
+    }
 
-	private final int byteArraySize;
-	private final int elementSize;
-	private final int headerSize;
-	private final int count;
-	private final byte[] bytes;
-	private final ParsedAddress address;
-	private final Object value;
-	private static final byte[] INVALID_BYTES = {};
-	private DataTypeFactory factory;
+    private final int byteArraySize;
+    private final int elementSize;
+    private final int headerSize;
+    private final int count;
+    private final byte[] bytes;
+    private final ParsedAddress address;
+    private final Object value;
+    private static final byte[] INVALID_BYTES = {};
+    private DataTypeFactory factory;
 
-	@Before
-	public void setup() {
-		MockitoAnnotations.initMocks(this);
+    @Before
+    public void setup()
+    {
+        MockitoAnnotations.initMocks(this);
 
-		factory = new DataTypeFactory();
-	}
+        factory = new DataTypeFactory();
+    }
 
-	@Test
-	public void getElementSize_None_Parameterized() {
-		assertEquals(elementSize, factory.getElementSize(address));
-	}
+    @Test
+    public void getElementSize_None_Parameterized()
+    {
+        assertEquals(elementSize, factory.getElementSize(address));
+    }
 
-	@Test
-	public void getHeaderSize_None_Parameterized() {
-		assertEquals(headerSize, factory.getHeaderSize(address));
-	}
+    @Test
+    public void getHeaderSize_None_Parameterized()
+    {
+        assertEquals(headerSize, factory.getHeaderSize(address));
+    }
 
-	@Test
-	public void fromBytes_ValidBytes_ReturnsCorrectValue() {
-		if (count > 1) {
-			assertArrayEquals((Object[]) value,
-					(Object[]) factory.fromBytes(bytes, address));
-		} else {
-			assertEquals(value, factory.fromBytes(bytes, address));
-		}
-	}
+    @Test
+    public void fromBytes_ValidBytes_ReturnsCorrectValue()
+    {
+        if ( count > 1 )
+        {
+            assertArrayEquals((Object[]) value,
+                              (Object[]) factory.fromBytes(bytes, address));
+        }
+        else
+        {
+            assertEquals(value, factory.fromBytes(bytes, address));
+        }
+    }
 
-	@Test(expected = DataTypeException.class)
-	public void fromBytes_InvalidBytes_ThrowsException() {
-		factory.fromBytes(INVALID_BYTES, address);
-	}
+    @Test(expected = DataTypeException.class)
+    public void fromBytes_InvalidBytes_ThrowsException()
+    {
+        factory.fromBytes(INVALID_BYTES, address);
+    }
 
-	@Test
-	public void toBytes_ValidValue_ReturnsCorrectBytes() {
-		byte[] result = factory.toBytes(value, address);
-		assertEquals(byteArraySize, result.length);
+    @Test
+    public void toBytes_ValidValue_ReturnsCorrectBytes()
+    {
+        byte[] result = factory.toBytes(value, address);
+        assertEquals(byteArraySize, result.length);
 
-		for (int i = 0; i < bytes.length; i++) {
-			assertEquals("Position [" + i + "]: " + bytes[i] + " != "
-					+ result[i], bytes[i], result[i]);
-		}
-	}
+        for ( int i = 0; i < bytes.length; i++ )
+        {
+            assertEquals("Position [" + i + "]: " + bytes[i] + " != "
+                + result[i], bytes[i], result[i]);
+        }
+    }
 }
